+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/process/AbstractPhaseAction.java
@@ -51,6 +51,7 @@ public abstract class AbstractPhaseAction implements PhaseAction {
 
     @Override
     public void execute() {
+        if (shouldSkip()) return;
         try {
             PhaseDriver.getContext().setPhase(getPhase());
             doExecute();